# vulkan: optimize iq1 coopmat2 dequant functions